Move vote_program out of src/
