+++ Cargo.toml
@@ -110,6 +110,7 @@ solana-metrics = { path = "metrics", version = "0.11.0" }
 solana-noop = { path = "programs/native/noop", version = "0.11.0" }
 solana-sdk = { path = "sdk", version = "0.11.0" }
 solana-system-program = { path = "programs/native/system", version = "0.11.0" }
+solana-vote-program = { path = "programs/native/vote", version = "0.11.0" }
 solana-vote-signer = { path = "vote-signer", version = "0.0.1" }
 sys-info = "0.5.6"
 tokio = "0.1"
