+++ programs/native/vote/Cargo.toml
@@ -1,21 +1,20 @@
 [package]
-name = "solana-sdk"
+name = "solana-vote-program"
 version = "0.11.0"
-description = "Solana SDK"
+description = "Solana vote program"
 authors = ["Solana Maintainers <maintainers@solana.com>"]
 repository = "https://github.com/solana-labs/solana"
 license = "Apache-2.0"
 
 [dependencies]
 bincode = "1.0.0"
-byteorder = "1.2.1"
-bs58 = "0.2.0"
-generic-array = { version = "0.12.0", default-features = false, features = ["serde"] }
+env_logger = "0.6.0"
 log = "0.4.2"
-ring = "0.13.2"
-sha2 = "0.8.0"
 serde = "1.0.27"
 serde_derive = "1.0.27"
-serde_json = "1.0.10"
-untrusted = "0.6.2"
+solana-sdk = { path = "../../../sdk", version = "0.11.0" }
+
+[lib]
+name = "solana_vote_program"
+crate-type = ["cdylib"]
