+++ src/bin/fullnode.rs
@@ -17,9 +17,9 @@ use solana::leader_scheduler::LeaderScheduler;
 use solana::logger;
 use solana::netutil::find_available_port_in_range;
 use solana::thin_client::poll_gossip_for_leader;
-use solana::vote_program::VoteProgram;
 use solana::vote_transaction::VoteTransaction;
 use solana_sdk::signature::{Keypair, KeypairUtil};
+use solana_sdk::vote_program::VoteProgram;
 use std::fs::File;
 use std::net::{Ipv4Addr, SocketAddr};
 use std::process::exit;
