+++ src/lib.rs
@@ -78,7 +78,6 @@ pub mod thin_client;
 pub mod tpu;
 pub mod tpu_forwarder;
 pub mod tvu;
-pub mod vote_program;
 pub mod vote_stage;
 pub mod vote_transaction;
 pub mod wallet;
